\name{Model3DCMCC-package}
\alias{Model3DCMCC-package}
\alias{Model3DCMCC}
\docType{package}
\title{
\packageTitle{Model3DCMCC}
}
\description{
\packageDescription{Model3DCMCC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Model3DCMCC}
\packageIndices{Model3DCMCC}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Model3DCMCC}

Maintainer: \packageMaintainer{Model3DCMCC}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
